as6502_note should take a ulong for lineNumber, since currentLineNum is also ulong
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@436 96dfbd92-d197-e211-9ca9-00110a534b34

+++ as6502/error.c
@@ -41,7 +41,7 @@ void as6502_warn(const char *reason) {
 	}
 }
 
-void as6502_note(int lineNumber, const char *reason, ...) {
+void as6502_note(unsigned long lineNumber, const char *reason, ...) {
 	va_list ap;
 	va_start(ap, reason);
